+++ src/renderer/App.tsx
@@ -11,8 +11,6 @@ import { getLogger } from '../shared/logger'
 import { BackendRemote } from './backend-com'
 import { runPostponedFunctions } from './onready'
 import { I18nContext } from './contexts/I18nContext'
-import { DialogContextProvider } from './contexts/DialogContext'
-import { ContextMenuProvider } from './contexts/ContextMenuContext'
 
 export default function App(_props: any) {
   const [localeData, setLocaleData] = useState<LocaleData | null>(null)
@@ -77,11 +75,7 @@ export default function App(_props: any) {
     <CrashScreen>
       <ThemeContextWrapper>
         <I18nContext.Provider value={window.static_translate}>
-          <ContextMenuProvider>
-            <DialogContextProvider>
-              <ScreenController />
-            </DialogContextProvider>
-          </ContextMenuProvider>
+          <ScreenController />
         </I18nContext.Provider>
       </ThemeContextWrapper>
     </CrashScreen>