+++ src/renderer/components/dialogs/FullscreenMedia.tsx
@@ -10,12 +10,11 @@ import { getLogger } from '../../../shared/logger'
 import { gitHubIssuesUrl } from '../../../shared/constants'
 import { useInitEffect } from '../helpers/hooks'
 import { preventDefault } from '../../../shared/util'
-import { jumpToMessage } from '../helpers/ChatMethods'
 import { BackendRemote, onDCEvent, Type } from '../../backend-com'
 import { selectedAccountId } from '../../ScreenController'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
-import useDialog from '../../hooks/useDialog'
 import useContextMenu from '../../hooks/useContextMenu'
+import useMessage from '../../hooks/chat/useMessage'
 
 import type { DialogProps } from '../../contexts/DialogContext'
 
@@ -35,7 +34,7 @@ type Props = {
 export default function FullscreenMedia(props: Props & DialogProps) {
   const accountId = selectedAccountId()
   const tx = useTranslationFunction()
-  const { openDialog } = useDialog()
+  const { jumpToMessage } = useMessage()
   const { onClose } = props
 
   const [msg, setMsg] = useState(props.msg)
@@ -105,7 +104,7 @@ export default function FullscreenMedia(props: Props & DialogProps) {
     {
       label: tx('show_in_chat'),
       action: () => {
-        jumpToMessage(msg.id)
+        jumpToMessage(accountId, msg.id)
         onClose()
       },
     },
@@ -147,10 +146,7 @@ export default function FullscreenMedia(props: Props & DialogProps) {
         <p>mimeType is "{fileMime}"</p>
         <p>
           Please report this bug on{' '}
-          <a
-            href='#'
-            onClick={() => runtime.openLink(openDialog, gitHubIssuesUrl)}
-          >
+          <a href='#' onClick={() => runtime.openLink(gitHubIssuesUrl)}>
             github
           </a>
         </p>
@@ -167,10 +163,7 @@ export default function FullscreenMedia(props: Props & DialogProps) {
         </p>
         <p>
           Please report this bug on{' '}
-          <a
-            href='#'
-            onClick={() => runtime.openLink(openDialog, gitHubIssuesUrl)}
-          >
+          <a href='#' onClick={() => runtime.openLink(gitHubIssuesUrl)}>
             github
           </a>
         </p>