+++ src/renderer/components/screens/CrashScreen.tsx
@@ -44,7 +44,6 @@ export class CrashScreen extends React.Component {
               href='#'
               onClick={_ =>
                 runtime.openLink(
-                  this.context.openDialog,
                   'https://github.com/deltachat/deltachat-desktop/issues'
                 )
               }