+++ src/renderer/hooks/usePrevious.ts
@@ -1,0 +1,18 @@
+import { useEffect, useRef } from 'react'
+
+/**
+ * Helper hook to look at the previous prop- or state value
+ * during the render cycle.
+ *
+ * Read more about this pattern here:
+ *
+ * - https://legacy.reactjs.org/docs/hooks-faq.html#how-to-get-the-previous-props-or-state
+ * - https://stackoverflow.com/questions/53446020/how-to-compare-oldvalues-and-newvalues-on-react-hooks-useeffect
+ */
+export default function usePrevious(value: any) {
+  const ref = useRef()
+  useEffect(() => {
+    ref.current = value
+  })
+  return ref.current
+}